add new class for avatar
